organizing chart display functions into separate file
